refactor ui tests.

+++ ui/src/actions/FriendsActions.js
@@ -10,32 +10,31 @@ export function getFriends() {
 }
 
 export function addFriend(name, sex) {
-  return {
-    type: types.ADD_FRIEND,
-    payload: axios.post(friends, { name, sex, starred: false })
+  if (name && sex){
+    return {
+      type: types.ADD_FRIEND,
+      payload: axios.post(friends, { name, sex, starred: false })
+    }
   }
+  return null
 }
 
 export function deleteFriend(id) {
-  return {
-    type: types.DELETE_FRIEND,
-    payload: axios.delete(`${friends}/${id}`)
-  }
- return dispatch => {
-   axios.delete(`${friends}/${id}`).then(
-    (payload) => {
-      dispatch({
-        type: types.DELETE_FRIEND,
-        payload,
-      })
+  if (id){
+    return {
+      type: types.DELETE_FRIEND,
+      payload: axios.delete(`${friends}/${id}`)
     }
-   )
- }
+  }
+  return null;
 }
 
 export function starFriend(id, starred) {
-  return {
-    type: types.STAR_FRIEND,
-    payload: axios.put(`${friends}/${id}`, { starred })
+  if(id){
+    return {
+      type: types.STAR_FRIEND,
+      payload: axios.put(`${friends}/${id}`, { starred })
+    }
   }
+  return null;
 }
